requirements.txt
beautifulsoup4==4.12.3
cachetools==5.5.2
certifi==2026.2.25
charset-normalizer==3.4.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
greenlet==3.3.2
gspread==6.0.2
httplib2==0.31.2
idna==3.11
oauth2client==4.1.3
oauthlib==3.3.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyparsing==3.3.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
StrEnum==0.4.15
typing_extensions==4.15.0
urllib3==2.6.3
